#1 Thanks, @newlukai , you're right. In the database model, carbs_per_100g is double. Consequently I will also change calories_per_100g, as this should also be double.

+++ app/src/main/java/info/rueth/fpucalculator/domain/repository/DatabaseJsonImportService.java
@@ -174,7 +174,7 @@ public class DatabaseJsonImportService extends ImportExportService {
                     food.setCaloriesPer100g(reader.nextInt());
                     break;
                 case DatabaseJsonExportService.CARBS_PER_100G:
-                    food.setCarbsPer100g(reader.nextInt());
+                    food.setCarbsPer100g(reader.nextDouble());
                     break;
                 case DatabaseJsonExportService.AMOUNT_SMALL:
                     food.setAmountSmall(reader.nextInt());
